requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
setuptools>=68.0.0